requirements.txt
apache-sedona==1.4.1
attrs==23.1.0
certifi==2023.7.22
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
delta-spark==2.4.0
Fiona==1.9.4.post1
geopandas==0.13.2
importlib-metadata==6.8.0
iniconfig==2.0.0
numpy==1.25.2
packaging==23.1
pandas==2.0.3
pluggy==1.2.0
py4j==0.10.9.7
pyproj==3.6.0
pyspark==3.4.0
pytest==7.4.0
python-dateutil==2.8.2
pytz==2023.3
shapely==2.0.1
six==1.16.0
termcolor==2.3.0
tzdata==2023.3
zipp==3.16.2
